+++ app/models/rooms.rb
@@ -1,2 +1,0 @@
-class Rooms < ApplicationRecord
-end