+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
   - [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Create a room PR#3 by @diabolo
 - Add changelog PR#2 by @diabolo